pyproject.toml
[project]
name = "rag-invoice-chat"
version = "0.1.0"
description = "RAG-style Chat Agent for Invoice Data"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pydantic>=2.11.5",
    "pydantic-settings>=2.9.1",
    "structlog>=25.4.0",
    "openpyxl",
    "pandas",
    "python-dotenv>=1.0.0",
    "openai>=1.0.0",
    "streamlit>=1.40.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.11.0",
    "pre-commit>=3.0.0"
]

[tool.mypy]
python_version = "3.13"
warn_unused_configs = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_return_any = true
warn_unreachable = true
strict = true
disallow_untyped_decorators = false
disallow_subclassing_any = false

[tool.ruff]
target-version = "py313"

[tool.ruff.lint]
ignore = ["D203", "D212", "D104", "PLR0913"]
select = [
    "E",     # pycodestyle
    "F",     # pyflakes
    "W",     # pycodestyle
    "I",     # isort
    "D",     # pydocstyle
    "UP",    # pyupgrade
    "N",     # pep8-naming
    "YTT",   # flake8-2020
    "ANN",   # flake8-annotations
    "ASYNC", # flake8-async
    "S",     # flake8-bandit
    "FBT",   # flake8-boolean-trap
    "B",     # flake8-bugbear
    "A",     # flake8-builtins
    "C4",    # flake8-comprehensions
    "DTZ",   # flake8-datetimez
    "T10",   # flake8-debugger
    "EM",    # flake8-errmsg
    "PIE",   # flake8-pie
    "T20",   # flake8-print
    "PT",    # flake8-pytest-style
    "Q",     # flake8-quotes
    "RSE",   # flake8-raise
    "RET",   # flake8-return
    "SLF",   # flake8-self
    "SIM",   # flake8-simplify
    "ARG",   # flake8-unused-arguments
    "PTH",   # flake8-use-pathlib
    "PL",    # pylint
    "TRY",   # tryceratops
    "FLY",   # flynt
    "NPY",   # NumPy-specific rules
    "PERF",  # Perflint
    "FURB",  # refurb
    "RUF",   # Ruff-specific rules
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]
